+++ sdk/cosmos/azure_data_cosmos/src/clients/cosmos_client.rs
@@ -54,7 +54,8 @@ impl CosmosClient {
         options: Option<CosmosClientOptions>,
     ) -> azure_core::Result<Self> {
         let options = options.unwrap_or_default();
-        let mut client_options = options.client_options;
+        let endpoint: Url = endpoint.parse()?;
+        let mut client_options = options.client_options.clone();
         client_options.retry = RetryOptions::none();
 
         let pipeline_core = azure_core::http::Pipeline::new(
@@ -68,16 +69,15 @@ impl CosmosClient {
             None,
         );
 
-        let global_endpoint_manager = GlobalEndpointManager::new(
-            endpoint.parse()?,
-            options.application_preferred_regions,
-            pipeline_core.clone(),
-        );
+        let preferred_regions = options.application_preferred_regions.clone();
+        let global_endpoint_manager =
+            GlobalEndpointManager::new(endpoint.clone(), preferred_regions, pipeline_core.clone());
 
         let pipeline = Arc::new(CosmosPipeline::new(
-            endpoint.parse()?,
+            endpoint,
             pipeline_core,
-            global_endpoint_manager.clone(),
+            global_endpoint_manager,
+            options,
         ));
 
         Ok(Self {
@@ -110,7 +110,9 @@ impl CosmosClient {
         options: Option<CosmosClientOptions>,
     ) -> azure_core::Result<Self> {
         let options = options.unwrap_or_default();
-        let mut client_options = options.client_options;
+        let endpoint: Url = endpoint.parse()?;
+
+        let mut client_options = options.client_options.clone();
         client_options.retry = RetryOptions::none();
 
         let pipeline_core = azure_core::http::Pipeline::new(
@@ -122,16 +124,15 @@ impl CosmosClient {
             None,
         );
 
-        let global_endpoint_manager = GlobalEndpointManager::new(
-            endpoint.parse()?,
-            options.application_preferred_regions,
-            pipeline_core.clone(),
-        );
+        let preferred_regions = options.application_preferred_regions.clone();
+        let global_endpoint_manager =
+            GlobalEndpointManager::new(endpoint.clone(), preferred_regions, pipeline_core.clone());
 
         let pipeline = Arc::new(CosmosPipeline::new(
-            endpoint.parse()?,
+            endpoint,
             pipeline_core,
-            global_endpoint_manager.clone(),
+            global_endpoint_manager,
+            options,
         ));
 
         Ok(Self {
@@ -247,7 +248,7 @@ impl CosmosClient {
 
         let cosmos_request =
             CosmosRequest::builder(OperationType::Create, self.databases_link.clone())
-                .headers(&options.throughput)
+                .request_headers(&options.throughput)
                 .json(&RequestBody { id })
                 .build()?;