+++ sdk/core/azure_core/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features Added
 
+- Added `Body::take()`.
 - Added `continuation_token` to `PagerOptions`.
 - Added extensible request authorization and authentication challenge handling to `BearerTokenAuthorizationPolicy`.
   - `OnRequest`, `OnChallenge`, and `Authorizer` traits define callbacks for these features.
@@ -15,12 +16,18 @@
 ### Breaking Changes
 
 - Added type parameter `C` to `Pager` declaration, defaulting to `Url` so it can be elided in most existing declarations.
+- Added type parameter `C` to `Poller` declaration, defaulting to `Url` so it can be elided in most existing declarations.
 - Changed `Pager::from_callback` to take a `PagerOptions` as the second parameter rather than a `Context` parameter.
 - Changed `Pager::from_callback` to `Pager::new` which now requires the caller to return a `Pin<Box<dyn Future>>`.
+- Changed `Poller::from_callback` to `Poller::new` which now requires the caller to return a `Pin<Box<dyn Future>>`.
+- Changed `PollerResult::next` field to `PollerResult::continuation_token`.
+- Changed `PollerState` generic parameter from `N` to `C`.
 - Moved `BearerTokenAuthorizationPolicy` into `azure_core::http::policies::auth`.
 - Removed `ItemIterator::with_continuation_token()`. Pass a continuation token to `PagerOptions::continuation_token` instead.
 - Removed `PageIterator::with_continuation_token()`. Pass a continuation token to `PagerOptions::continuation_token` instead.
 - Removed `Pager::from_stream`.
+- Removed `Poller::from_stream`.
+- Renamed `BoxedFuture` to `PagerResultFuture` in `pager` module.
 
 ### Bugs Fixed