+++ sdk/storage/azure_storage_blob/CHANGELOG.md
@@ -10,16 +10,15 @@
 - Added support for `set_immutability_policy` to `BlobClient`.
 - Added support for `delete_immutability_policy` to `BlobClient`.
 - Added support for `undelete` to `BlobClient`.
+- Added snapshot and versioning support for blobs with convenience methods `with_snapshot` and `with_version` to `BlobClient`.
 
 ### Breaking Changes
 
 - Changed conversion implementation from `BlobTags` to `HashMap<String, String>` from `TryFrom` to `From`.
 - Added `continuation_token` to `PagerOptions` for methods that return a `Pager`.
 - Renamed `content_length` to `size` for `PageBlobClient`'s `create()` method.
-
-### Breaking Changes
-
 - Removed `Pager::with_continuation_token()` for methods that return a `Pager`.
+- Changed `BlobClient`'s `set_tags` parameter `tags` type from `HashMap<String, String>` to `RequestContent<BlobTags, XmlFormat>`.
 
 ### Bugs Fixed